remove clusterlistfromgraph

+++ src/main/java/org/opentripplanner/graph_builder/module/time/EdgeData.java
@@ -27,7 +27,7 @@ public class EdgeData {
         this.clusterid = clusterid;
     }
 
-    public double getstartnodeid() {
+    public long  getstartnodeid() {
         return startnodeid;
     }
